+++ CHANGELOG.md
@@ -1,10 +1,34 @@
 
 ## Fixed:
 - `ImageSlideShow`
-- [BREAKING] Hide overlay widget class: `ImageSlideShow`. #42
+- Hide overlay when widget class dispose: `ImageSlideShow`. #83
+- 
+[Changes][1.0.5]
+
+## Added:
+- `pubspec`
+- reduce min sdk to: `2.18.0`.
+- 
+[Changes][1.0.4]
+
+## Added:
+- `pubspec`
+- reduce min sdk to: `2.12.0`.
+- 
+[Changes][1.0.3]
+
+## Fixed:
+- `ImageSlideShow`
+- Add material for class: `ImageSlideShow`.
+- 
+[Changes][1.0.2]
+
+## Fixed:
+- `ImageSlideShow`
+- Hide overlay widget class: `ImageSlideShow`. #42
 [Changes][1.0.1]
 
 ## Added:
 - `ImageSlideWidget` and `ImageSlideShow`
-
+- 
 [Changes][1.0.0]